+++ pytket/binders/circuit/unitid.cpp
@@ -85,6 +85,7 @@ void init_unitid(py::module &m) {
 
   py::class_<UnitID>(
       m, "UnitID", "A handle to a computational unit (e.g. qubit, bit)")
+      .def(py::init<>())
       .def("__eq__", &UnitID::operator==)
       .def("__lt__", &UnitID::operator<)
       .def("__repr__", &UnitID::repr)