+++ pytket/tests/transform_test.py
@@ -29,6 +29,7 @@ from pytket.passes import (  # type: ignore
     DefaultMappingPass,
     FullMappingPass,
     RoutingPass,
+    CustomRoutingPass,
     PlacementPass,
     CXMappingPass,
     auto_rebase_pass,
@@ -798,13 +799,18 @@ def test_RoutingPass() -> None:
     circ.CX(1, 3)
     circ.CX(1, 2)
     cu_0 = CompilationUnit(circ)
+    cu_1 = CompilationUnit(circ)
     placer = GraphPlacement(arc)
     p_pass = PlacementPass(placer)
     r_pass_0 = RoutingPass(arc)
+    r_pass_1 = CustomRoutingPass(arc, [LexiLabellingMethod(), LexiRouteRoutingMethod()])
     p_pass.apply(cu_0)
     r_pass_0.apply(cu_0)
+    p_pass.apply(cu_1)
+    r_pass_1.apply(cu_1)
     out_circ_0 = cu_0.circuit
     assert out_circ_0.valid_connectivity(arc, False, True)
+    assert out_circ_0 == cu_1.circuit
 
 
 def test_FullMappingPass() -> None: