+++ tket/src/Predicates/include/Predicates/PassLibrary.hpp
@@ -18,6 +18,7 @@
 
 namespace tket {
 
+const PassPtr &SynthesiseTK();
 const PassPtr &SynthesiseTket();
 const PassPtr &SynthesiseHQS();
 const PassPtr &SynthesiseOQC();
@@ -33,11 +34,16 @@ const PassPtr &DecomposeArbitrarilyControlledGates();
 const PassPtr &DecomposeMultiQubitsCX();
 const PassPtr &DecomposeSingleQubitsTK1();
 const PassPtr &DecomposeBoxes();
+
 /**
  * converts a circuit containing all possible gates to a circuit containing only
  * phase poly boxes + H gates (and measure + reset + collapse + barrier)
+ * @param min_size value for the minimal number of CX in each box, groups with
+ * less than min_size CX gates are not converted to a PhasePolyBox, default
+ * value is 0
+ * @return PassPtr to perform the conversion
  */
-const PassPtr &ComposePhasePolyBoxes();
+PassPtr ComposePhasePolyBoxes(unsigned min_size = 0);
 
 /** Squash sequences of single-qubit gates to TK1 gates. */
 const PassPtr &SquashTK1();