+++ tket/src/Transformations/Decomposition.cpp
@@ -38,6 +38,32 @@ static bool convert_to_zyz(Circuit &circ);
 static bool convert_to_xyx(Circuit &circ);
 
 /**
+ * Decompose all multi-qubit unitary gates into TK2 and single-qubit gates.
+ *
+ * This function does not decompose boxes.
+ */
+static bool convert_multiqs_TK2(Circuit &circ) {
+  bool success = false;
+  VertexList bin;
+  BGL_FORALL_VERTICES(v, circ.dag, DAG) {
+    Op_ptr op = circ.get_Op_ptr_from_Vertex(v);
+    OpType optype = op->get_type();
+    if (is_gate_type(optype) && !is_projective_type(optype) &&
+        op->n_qubits() >= 2 && (optype != OpType::TK2)) {
+      Circuit in_circ = TK2_circ_from_multiq(op);
+      Subcircuit sub = {
+          {circ.get_in_edges(v)}, {circ.get_all_out_edges(v)}, {v}};
+      bin.push_back(v);
+      circ.substitute(in_circ, sub, Circuit::VertexDeletion::No);
+      success = true;
+    }
+  }
+  circ.remove_vertices(
+      bin, Circuit::GraphRewiring::No, Circuit::VertexDeletion::Yes);
+  return success;
+}
+
+/**
  * Decompose all multi-qubit unitary gates into CX and single-qubit gates.
  *
  * This function does not decompose boxes.
@@ -129,6 +155,10 @@ static bool convert_to_xyx(Circuit &circ) {
   circ.remove_vertices(
       bin, Circuit::GraphRewiring::No, Circuit::VertexDeletion::Yes);
   return success;
+}
+
+Transform decompose_multi_qubits_TK2() {
+  return Transform(convert_multiqs_TK2);
 }
 
 Transform decompose_multi_qubits_CX() { return Transform(convert_multiqs_CX); }
@@ -745,8 +775,8 @@ Transform decomp_boxes() {
   return Transform([](Circuit &circ) { return circ.decompose_boxes(); });
 }
 
-Transform compose_phase_poly_boxes() {
-  return Transform([](Circuit &circ) {
+Transform compose_phase_poly_boxes(const unsigned min_size) {
+  return Transform([=](Circuit &circ) {
     // replace wireswaps with three CX
     while (circ.has_implicit_wireswaps()) {
       qubit_map_t perm = circ.implicit_qubit_permutation();
@@ -758,7 +788,7 @@ Transform compose_phase_poly_boxes() {
       }
     }
 
-    CircToPhasePolyConversion conv = CircToPhasePolyConversion(circ);
+    CircToPhasePolyConversion conv = CircToPhasePolyConversion(circ, min_size);
     conv.convert();
     circ = conv.get_circuit();
     return true;